Move update methods from `TaskService` to `TaskUpdateManager`
